Refs: #11666 added TestUtils/DatabaseUtils for help in test, modified tests to use it
