+++ test/unittest/StatisticsBackend/StatisticsBackendTests.cpp
@@ -20,6 +20,8 @@
 
 #include "gtest/gtest.h"
 
+#include <DatabaseUtils.hpp>
+
 using namespace eprosima::statistics_backend;
 using namespace eprosima::statistics_backend::database;
 
@@ -28,159 +30,20 @@ using namespace eprosima::statistics_backend::database;
  *
  * \c get_type Returns the entity kind of a given id.
  *
- * This fixture populates a database with several entities and several relations among them.
- * Since the tests need to know these relations, each entity is given a unique identifier within the fixture,
- * and the fixture keeps all these entities on a map,
- * for the test to be able to get them and check the result of the execution.
- * This identifier is independent of the EntityId given by the database,
- * and is totally under control of the fixture and the tests.
- * This is important to keep the tests stable even if the Database implementation to assign an EntityId changes.
- *
- * The following Json object describes the populated database objects and their given unique identifiers:
- *
- * \code
- * {
- *  "host1": {
- *     "ID": 1,
- *     "users": []
- *   },
- *   "host": {
- *     "ID": 2,
- *     "users": ["user1", "user"]
- *   },
- *   "user1": {
- *     "ID": 3,
- *     "processes": []
- *   },
- *   "user": {
- *     "ID": 4,
- *     "processes": ["process1", "process"]
- *   },
- *   "process1": {
- *     "ID": 5,
- *     "participants": []
- *   },
- *   "process": {
- *     "ID": 6,
- *     "participants": ["participant1", "participant"]
- *   },
- *   "domain1": {
- *     "ID": 7,
- *     "participants": [],
- *     "topics": []
- *   },
- *   "domain": {
- *     "ID": 8,
- *     "participants": ["participant1", "participant"],
- *     "topics": ["topic1", "topic"]
- *   },
- *   "participant1": {
- *     "ID": 9,
- *     "datareaders": [],
- *     "datawriters": []
- *   },
- *   "participant": {
- *     "ID": 10,
- *     "datareaders": ["datareader1", "datareader"],
- *     "datawriters": ["datawriter1", "datawriter"]
- *   },
- *   "topic1": {
- *     "ID": 11,
- *     "datareaders": [],
- *     "datawriters": []
- *   },
- *   "topic": {
- *     "ID": 12,
- *     "datareaders": ["datareader1", "datareader"],
- *     "datawriters": ["datawriter1", "datawriter"]
- *   },
- *   "datareader1": {
- *     "ID": 13,
- *     "locators": ["reader_locator1"]
- *   }
- *   "datareader": {
- *     "ID": 14,
- *     "locators": ["reader_locator1", "reader_locator"]
- *   }
- *   "datawriter1": {
- *     "ID": 15,
- *     "locators": ["writer_locator1"]
- *   }
- *   "datawriter": {
- *     "ID": 16,
- *     "locators": ["writer_locator1", "writer_locator"]
- *   }
- *   "reader_locator1": {
- *     "ID": 17
- *   }
- *   "reader_locator": {
- *     "ID": 18
- *   }
- *   "writer_locator1": {
- *     "ID": 19
- *   }
- *   "writer_locator": {
- *     "ID": 20
- *   }
- * }
- * \endcode
  */
 class statistics_backend_tests : public ::testing::Test
 {
 public:
 
+    typedef uint32_t TestId;
+
     void SetUp()
     {
-        populate_database();
+        entities = PopulateDatabase::populate_database(db);
     }
 
-    void populate_database()
-    {
-        host = std::make_shared<Host>("host");
-        db.insert(host);
-
-        user = std::make_shared<User>("user", host);
-        db.insert(user);
-
-        process = std::make_shared<Process>("process", "345", user);
-        db.insert(process);
-
-        domain = std::make_shared<Domain>("domain");
-        db.insert(domain);
-
-        participant = std::make_shared<DomainParticipant>("participant", "qos", "5.6.7.8", nullptr, domain);
-        db.insert(participant);
-        db.link_participant_with_process(participant->id, process->id);
-
-        topic = std::make_shared<Topic>("topic", "type", domain);
-        db.insert(topic);
-
-        datareader = std::make_shared<DataReader>("datareader", "qos", "15.16.17.18", participant, topic);
-        reader_locator = std::make_shared<Locator>("reader_locator");
-        reader_locator->id = db.generate_entity_id();
-        datareader->locators[reader_locator->id] = reader_locator;
-        datareader->locators[reader_locator->id] = reader_locator;
-        db.insert(datareader);
-
-        datawriter = std::make_shared<DataWriter>("datawriter", "qos", "25.26.27.28", participant, topic);
-        writer_locator = std::make_shared<Locator>("writer_locator");
-        writer_locator->id = db.generate_entity_id();
-        datawriter->locators[writer_locator->id] = writer_locator;
-        datawriter->locators[writer_locator->id] = writer_locator;
-        db.insert(datawriter);
-    }
-
-    Database db;
-    std::shared_ptr<Host> host;
-    std::shared_ptr<User> user;
-    std::shared_ptr<Process> process;
-    std::shared_ptr<Domain> domain;
-    std::shared_ptr<DomainParticipant> participant;
-    std::shared_ptr<Topic> topic;
-    std::shared_ptr<DataReader> datareader;
-    std::shared_ptr<Locator> reader_locator;
-    std::shared_ptr<DataWriter> datawriter;
-    std::shared_ptr<Locator> writer_locator;
+    DataBaseTest db;
+    std::map<TestId, std::shared_ptr<const Entity>> entities;
 };
 
 class StatisticsBackendTest : public StatisticsBackend
@@ -199,16 +62,76 @@ public:
 TEST_F(statistics_backend_tests, get_type)
 {
     StatisticsBackendTest::set_database(&db);
-    ASSERT_EQ(StatisticsBackendTest::get_type(host->id), host->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(user->id), user->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(process->id), process->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(domain->id), domain->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(participant->id), participant->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(topic->id), topic->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(datareader->id), datareader->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(reader_locator->id), reader_locator->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(datawriter->id), datawriter->kind);
-    ASSERT_EQ(StatisticsBackendTest::get_type(writer_locator->id), writer_locator->kind);
+
+    for (auto pair : db.hosts())
+    {
+        auto entity = pair.second;
+        ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+    }
+
+    for (auto pair : db.users())
+    {
+        auto entity = pair.second;
+        ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+    }
+
+    for (auto pair : db.processes())
+    {
+        auto entity = pair.second;
+        ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+    }
+
+    for (auto pair : db.domains())
+    {
+        auto entity = pair.second;
+        ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+    }
+
+    for (auto domainPair : db.participants())
+    {
+        auto domainEntities = domainPair.second;
+        for (auto pair : domainEntities)
+        {
+            auto entity = pair.second;
+            ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+        }
+    }
+
+    for (auto domainPair : db.topics())
+    {
+        auto domainEntities = domainPair.second;
+        for (auto pair : domainEntities)
+        {
+            auto entity = pair.second;
+            ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+        }
+    }
+
+    for (auto domainPair : db.get_dds_endpoints<DataReader>())
+    {
+        auto domainEntities = domainPair.second;
+        for (auto pair : domainEntities)
+        {
+            auto entity = pair.second;
+            ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+        }
+    }
+
+    for (auto domainPair : db.get_dds_endpoints<DataWriter>())
+    {
+        auto domainEntities = domainPair.second;
+        for (auto pair : domainEntities)
+        {
+            auto entity = pair.second;
+            ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+        }
+    }
+
+    for (auto pair : db.locators())
+    {
+        auto entity = pair.second;
+        ASSERT_EQ(StatisticsBackendTest::get_type(entity->id), entity->kind);
+    }
 
     ASSERT_THROW(StatisticsBackendTest::get_type(EntityId::all()), BadParameter);
     ASSERT_THROW(StatisticsBackendTest::get_type(EntityId::invalid()), BadParameter);
